// frontend/src/app/core/api/generated/api/products.service.ts
export interface FindAll4RequestParams {
    page?: number;
    size?: number;
    sku?: string;
    name?: string;
    active?: boolean;
}